+++ packages/core/mashroom-utils/src/ts_node_utils.ts
@@ -1,0 +1,5 @@
+
+export const withinTsNode = () => {
+    // @ts-ignore
+    return !!process[Symbol.for('ts-node.register.instance')];
+};