+++ packages/core/mashroom/src/logging/delegates/log4js/MashroomLoggerDelegateLog4js.ts
@@ -2,6 +2,7 @@
 import os from 'os';
 import {existsSync} from 'fs';
 import log4js from 'log4js';
+import {withinTsNode} from '@mashroom/mashroom-utils/lib/ts_node_utils';
 import defaultConfig from './log4js_default_config';
 
 import type {LogLevel} from '../../../../type-definitions';
@@ -11,12 +12,16 @@ const ENVIRONMENT = process.env.NODE_ENV || 'development';
 const HOSTNAME = os.hostname() || 'localhost';
 
 const CONFIG_FILES = [
+    `log4js.${HOSTNAME}.${ENVIRONMENT}.ts`,
     `log4js.${HOSTNAME}.${ENVIRONMENT}.js`,
     `log4js.${HOSTNAME}.${ENVIRONMENT}.json`,
+    `log4js.${HOSTNAME}.ts`,
     `log4js.${HOSTNAME}.js`,
     `log4js.${HOSTNAME}.json`,
+    `log4js.${ENVIRONMENT}.ts`,
     `log4js.${ENVIRONMENT}.js`,
     `log4js.${ENVIRONMENT}.json`,
+    'log4js.ts',
     'log4js.js',
     'log4js.json',
 ];
@@ -28,18 +33,23 @@ export default class MashroomLoggerDelegateLog4js implements MashroomLoggerDeleg
 
     async init(serverRootPath: string) {
         try {
-            const configFiles = CONFIG_FILES.map((name) => `${serverRootPath}/${name}`);
+            let possibleConfigFiles = CONFIG_FILES;
+            if (!withinTsNode()) {
+                possibleConfigFiles = possibleConfigFiles.filter((p) => !p.endsWith('.ts'));
+            }
+            const configFiles = possibleConfigFiles.map((name) => `${serverRootPath}/${name}`);
             const configFile = configFiles.find((file) => existsSync(file));
 
             if (configFile) {
                 // eslint-disable-next-line @typescript-eslint/no-var-requires
-                const logConfig = require(configFile);
+                const logConfigModule = require(configFile);
+                const logConfig = logConfigModule.default ?? logConfigModule;
                 log4js.configure(logConfig);
-                log4js.getLogger().info('Considering log config files: ', CONFIG_FILES);
+                log4js.getLogger().info('Considering log config files (take the first match): ', possibleConfigFiles);
                 log4js.getLogger().info(`log4js configured from: ${configFile}`);
             } else {
                 log4js.configure(defaultConfig);
-                log4js.getLogger().info('Considering log config files: ', CONFIG_FILES);
+                log4js.getLogger().info('Considering log config files: ', possibleConfigFiles);
                 log4js.getLogger().warn('No log config files found, using default config');
             }
         } catch (error) {