+++ packages/plugin-packages/mashroom-security-provider-simple/src/MashroomSimpleSecurityProvider.ts
@@ -139,7 +139,8 @@ export default class MashroomSimpleSecurityProvider implements MashroomSecurityP
         let userStore: UserStore;
         if (fs.existsSync(this._userStorePath)) {
             // eslint-disable-next-line @typescript-eslint/no-var-requires
-            const userData = require(this._userStorePath);
+            const userDataModule = require(this._userStorePath);
+            const userData = userDataModule.default ?? userDataModule;
             userStore = Array.isArray(userData) ? userData : (userData.users || []);
             this._createRoleDefinitions(userStore, request, logger);
         } else {