+++ packages/test/test-server7/remotePortalApps.ts
@@ -1,0 +1,11 @@
+
+import type {MashroomPortalRemoteApps} from '@mashroom/mashroom-json-schemas/type-definitions';
+
+const remotePortalApps: MashroomPortalRemoteApps = {
+    remotePortalApps: [
+        'http://demo-ssr-remote-app.mashroom-server.com'
+    ]
+};
+
+export default remotePortalApps;
+