+++ packages/test/test-server7/test-plugin1/mashroom.ts
@@ -1,0 +1,18 @@
+
+import type {MashroomPlugins} from '@mashroom/mashroom-json-schemas/type-definitions';
+
+const plugins: MashroomPlugins = {
+    plugins: [
+        {
+            name: 'Test Webapp Plugin with TS config',
+            type: 'web-app',
+            bootstrap: './bootstrap.ts',
+            defaultConfig: {
+                path: '/test-webapp-ts-config',
+                name: 'World'
+            }
+        }
+    ]
+}
+
+export default plugins;