+++ packages/test/test-server7/topicACL.ts
@@ -1,0 +1,10 @@
+
+import type {MashroomSecurityTopicAcl} from '@mashroom/mashroom-json-schemas/type-definitions';
+
+const topicACL: MashroomSecurityTopicAcl = {
+    'external2/#': {
+        allow: ['Administrator']
+    }
+};
+
+export default topicACL;